+++ unreal-editor-mcp/plugin/UnrealMCP/Source/UnrealMCP/Private/UnrealMCPApiProbe.cpp
@@ -6,6 +6,7 @@
 #include "BlueprintActionDatabase.h"
 #include "BlueprintActionFilter.h"
 #include "BlueprintFunctionNodeSpawner.h"
+#include "BlueprintEventNodeSpawner.h"
 #include "BlueprintVariableNodeSpawner.h"
 #include "EdGraphSchema_K2.h"
 #include "FileHelpers.h"
@@ -17,7 +18,11 @@
 #include "K2Node_FunctionEntry.h"
 #include "K2Node_FunctionResult.h"
 #include "K2Node_CustomEvent.h"
+#include "K2Node_DynamicCast.h"
+#include "K2Node_IfThenElse.h"
+#include "K2Node_EnumLiteral.h"
 #include "K2Node_MacroInstance.h"
+#include "K2Node_PromotableOperator.h"
 #include "K2Node_Tunnel.h"
 #include "ScopedTransaction.h"
 #include "SubobjectDataSubsystem.h"
@@ -31,10 +36,15 @@ void RequirePublicTypes()
     static_assert(TIsDerivedFrom<UEdGraphSchema_K2, UEdGraphSchema>::Value);
     static_assert(TIsDerivedFrom<UBlueprintNodeSpawner, UObject>::Value);
     static_assert(TIsDerivedFrom<UBlueprintFunctionNodeSpawner, UBlueprintNodeSpawner>::Value);
+    static_assert(TIsDerivedFrom<UBlueprintEventNodeSpawner, UBlueprintNodeSpawner>::Value);
     static_assert(TIsDerivedFrom<UBlueprintVariableNodeSpawner, UBlueprintNodeSpawner>::Value);
     static_assert(TIsDerivedFrom<UK2Node_FunctionEntry, UK2Node>::Value);
     static_assert(TIsDerivedFrom<UK2Node_FunctionResult, UK2Node>::Value);
     static_assert(TIsDerivedFrom<UK2Node_CustomEvent, UK2Node>::Value);
+    static_assert(TIsDerivedFrom<UK2Node_DynamicCast, UK2Node>::Value);
+    static_assert(TIsDerivedFrom<UK2Node_IfThenElse, UK2Node>::Value);
+    static_assert(TIsDerivedFrom<UK2Node_EnumLiteral, UK2Node>::Value);
+    static_assert(TIsDerivedFrom<UK2Node_PromotableOperator, UK2Node>::Value);
     static_assert(TIsDerivedFrom<UK2Node_MacroInstance, UK2Node>::Value);
     static_assert(TIsDerivedFrom<UK2Node_Tunnel, UK2Node>::Value);
     static_assert(TIsDerivedFrom<USubobjectDataSubsystem, UEngineSubsystem>::Value);